+++ phpshop/admpanel/exchange/admin_import.php
@@ -95,6 +95,9 @@ if ($GLOBALS['PHPShopBase']->codBase == 'utf-8')
 // Стоп лист
 $key_stop = array('password', 'wishlist', 'sort', 'yml_bid_array', 'status', 'files', 'datas', 'price_search', 'vid', 'name_rambler', 'servers', 'skin', 'skin_enabled', 'secure_groups', 'icon_description', 'title_enabled', 'title_shablon', 'descrip_shablon', 'descrip_enabled', 'productsgroup_check', 'productsgroup_product', 'keywords_enabled', 'keywords_shablon', 'rate_count', 'sort_cache', 'sort_cache_created_at', 'parent_title', 'menu', 'order_by', 'order_to', 'org_ras', 'org_bank', 'org_kor', 'org_bik', 'org_city', 'admin', 'org_fakt_adres');
 
+if(empty($subpath[2]))
+    $subpath[2]=null;
+
 switch ($subpath[2]) {
     case 'catalog':
         $PHPShopOrm = new PHPShopOrm($GLOBALS['SysValue']['base']['categories']);
@@ -380,7 +383,7 @@ function csv_update($data) {
             }
 
             // Полный путь к изображениями
-            if (isset($_POST['export_imgpath'])) {
+            if (!empty($_POST['export_imgpath']) and isset($_POST['export_imgpath'])) {
                 if (!empty($row['pic_small']))
                     $row['pic_small'] = '/UserFiles/Image/' . $row['pic_small'];
             }
@@ -388,10 +391,10 @@ function csv_update($data) {
             // Дополнительные изображения
             if (!empty($_POST['export_imgdelim']) and strstr($row['pic_big'], $_POST['export_imgdelim'])) {
                 $data_img = explode($_POST['export_imgdelim'], $row['pic_big']);
-            } else
+            } elseif(!empty($row['pic_big']))
                 $data_img[] = $row['pic_big'];
 
-            if (is_array($data_img)) {
+            if (!empty($data_img) and is_array($data_img)) {
 
                 // Получение ID товара по артикулу
                 if (empty($row['id']) and ! empty($row['uid'])) {
@@ -641,7 +644,7 @@ function actionSave() {
     }
 
     // Удалить настройки
-    if (is_array($_POST['exchanges_remove'])) {
+    if (!empty($_POST['exchanges_remove']) and is_array($_POST['exchanges_remove'])) {
         $PHPShopOrm = new PHPShopOrm($GLOBALS['SysValue']['base']['exchanges']);
         foreach ($_POST['exchanges_remove'] as $v)
             $data = $PHPShopOrm->delete(array('id' => '=' . intval($v)));
@@ -673,14 +676,14 @@ function actionSave() {
     // Память разделителей характеристик
     $memory = json_decode($_COOKIE['check_memory'], true);
     unset($memory[$_GET['path']]);
-    $memory[$_GET['path']]['export_sortdelim'] = $_POST['export_sortdelim'];
-    $memory[$_GET['path']]['export_sortsdelim'] = $_POST['export_sortsdelim'];
-    $memory[$_GET['path']]['export_imgdelim'] = $_POST['export_imgdelim'];
-    $memory[$_GET['path']]['export_imgpath'] = $_POST['export_imgpath'];
-    $memory[$_GET['path']]['export_uniq'] = $_POST['export_uniq'];
-    $memory[$_GET['path']]['export_action'] = $_POST['export_action'];
-    $memory[$_GET['path']]['export_delim'] = $_POST['export_delim'];
-    $memory[$_GET['path']]['export_imgproc'] = $_POST['export_imgproc'];
+    $memory[$_GET['path']]['export_sortdelim'] = @$_POST['export_sortdelim'];
+    $memory[$_GET['path']]['export_sortsdelim'] = @$_POST['export_sortsdelim'];
+    $memory[$_GET['path']]['export_imgdelim'] = @$_POST['export_imgdelim'];
+    $memory[$_GET['path']]['export_imgpath'] = @$_POST['export_imgpath'];
+    $memory[$_GET['path']]['export_uniq'] = @$_POST['export_uniq'];
+    $memory[$_GET['path']]['export_action'] = @$_POST['export_action'];
+    $memory[$_GET['path']]['export_delim'] = @$_POST['export_delim'];
+    $memory[$_GET['path']]['export_imgproc'] = @$_POST['export_imgproc'];
 
     if (is_array($memory))
         setcookie("check_memory", json_encode($memory), time() + 3600000, $GLOBALS['SysValue']['dir']['dir'] . '/phpshop/admpanel/');
@@ -942,16 +945,19 @@ function actionStart() {
 
     if (!empty($_COOKIE['check_memory'])) {
         $memory = json_decode($_COOKIE['check_memory'], true);
-        $export_sortdelim = $memory[$_GET['path']]['export_sortdelim'];
-        $export_sortsdelim = $memory[$_GET['path']]['export_sortsdelim'];
-        $export_imgvalue = $memory[$_GET['path']]['export_imgdelim'];
+        $export_sortdelim = @$memory[$_GET['path']]['export_sortdelim'];
+        $export_sortsdelim = @$memory[$_GET['path']]['export_sortsdelim'];
+        $export_imgvalue = @$memory[$_GET['path']]['export_imgdelim'];
+    }
+    else {
+        $memory=$export_sortdelim=$export_sortsdelim=$export_imgvalue=$exchanges_remove_value=null;
     }
 
-    $delim_value[] = array('Точка с запятой', ';', $memory[$_GET['path']]['export_delim']);
-    $delim_value[] = array('Запятая', ',', $memory[$_GET['path']]['export_delim']);
+    $delim_value[] = array('Точка с запятой', ';', @$memory[$_GET['path']]['export_delim']);
+    $delim_value[] = array('Запятая', ',', @$memory[$_GET['path']]['export_delim']);
 
-    $action_value[] = array('Обновление', 'update', $memory[$_GET['path']]['export_action']);
-    $action_value[] = array('Создание', 'insert', $memory[$_GET['path']]['export_action']);
+    $action_value[] = array('Обновление', 'update', @$memory[$_GET['path']]['export_action']);
+    $action_value[] = array('Создание', 'insert', @$memory[$_GET['path']]['export_action']);
 
     $delim_sortvalue[] = array('#', '#', $export_sortdelim);
     $delim_sortvalue[] = array('@', '@', $export_sortdelim);
@@ -973,42 +979,45 @@ function actionStart() {
     $key_value[] = array('Id или Артикул', 0, 'selected');
 
     // Закладка 1
-    $Tab1 = $PHPShopGUI->setField('Действие', $PHPShopGUI->setSelect('export_action', $action_value, 150, true)) .
+    $Tab1 = $PHPShopGUI->setField("Файл", $PHPShopGUI->setFile()).
+            $PHPShopGUI->setField('Действие', $PHPShopGUI->setSelect('export_action', $action_value, 150, true)) .
             $PHPShopGUI->setField('CSV-разделитель', $PHPShopGUI->setSelect('export_delim', $delim_value, 150, true)) .
             $PHPShopGUI->setField('Разделитель для характеристик', $PHPShopGUI->setSelect('export_sortdelim', $delim_sortvalue, 150), false, false, $class) .
             $PHPShopGUI->setField('Разделитель значений характеристик', $PHPShopGUI->setSelect('export_sortsdelim', $delim_sort, 150), false, false, $class) .
-            $PHPShopGUI->setField('Полный путь для изображений', $PHPShopGUI->setCheckbox('export_imgpath', 1, 'Включить', $memory[$_GET['path']]['export_imgpath']), 1, 'Добавляет к изображениям папку /UserFiles/Image/', $class) .
-            $PHPShopGUI->setField('Обработка изображений', $PHPShopGUI->setCheckbox('export_imgproc', 1, 'Включить', $memory[$_GET['path']]['export_imgproc']), 1, 'Создание тумбнейла и ватермарка', $class) .
+            $PHPShopGUI->setField('Полный путь для изображений', $PHPShopGUI->setCheckbox('export_imgpath', 1, null, @$memory[$_GET['path']]['export_imgpath']), 1, 'Добавляет к изображениям папку /UserFiles/Image/', $class) .
+            $PHPShopGUI->setField('Обработка изображений', $PHPShopGUI->setCheckbox('export_imgproc', 1, null, @$memory[$_GET['path']]['export_imgproc']), 1, 'Создание тумбнейла и ватермарка', $class) .
             $PHPShopGUI->setField('Разделитель для изображений', $PHPShopGUI->setSelect('export_imgdelim', $delim_imgvalue, 150), 1, 'Дополнительные изображения', $class) .
             $PHPShopGUI->setField('Кодировка текста', $PHPShopGUI->setSelect('export_code', $code_value, 150)) .
             $PHPShopGUI->setField('Ключ обновления', $PHPShopGUI->setSelect('export_key', $key_value, 150, false, false, true), 1, 'Изменение ключа обновления может привести к порче данных', $class) .
-            $PHPShopGUI->setField('Проверка уникальности', $PHPShopGUI->setCheckbox('export_uniq', 1, 'Включить', $memory[$_GET['path']]['export_uniq']), 1, 'Исключает дублирование данных при создании') .
-            $PHPShopGUI->setField("Файл", $PHPShopGUI->setFile());
+            $PHPShopGUI->setField('Проверка уникальности', $PHPShopGUI->setCheckbox('export_uniq', 1, null, @$memory[$_GET['path']]['export_uniq']), 1, 'Исключает дублирование данных при создании');
 
     // Закладка 2
-    $Tab2 = $PHPShopGUI->setField(array('Колонка 1', 'Колонка 2'), array($PHPShopGUI->setSelect('select_action[]', $select_value, 150, true), $PHPShopGUI->setSelect('select_action[]', $select_value, 150, true)), array(array(3, 2), array(2, 2)));
-    $Tab2 .= $PHPShopGUI->setField(array('Колонка 3', 'Колонка 4'), array($PHPShopGUI->setSelect('select_action[]', $select_value, 150, true), $PHPShopGUI->setSelect('select_action[]', $select_value, 150, true)), array(array(3, 2), array(2, 2)));
-    $Tab2 .= $PHPShopGUI->setField(array('Колонка 5', 'Колонка 6'), array($PHPShopGUI->setSelect('select_action[]', $select_value, 150, true), $PHPShopGUI->setSelect('select_action[]', $select_value, 150, true)), array(array(3, 2), array(2, 2)));
-    $Tab2 .= $PHPShopGUI->setField(array('Колонка 7', 'Колонка 8'), array($PHPShopGUI->setSelect('select_action[]', $select_value, 150, true), $PHPShopGUI->setSelect('select_action[]', $select_value, 150, true)), array(array(3, 2), array(2, 2)));
-    $Tab2 .= $PHPShopGUI->setField(array('Колонка 9', 'Колонка 10'), array($PHPShopGUI->setSelect('select_action[]', $select_value, 150, true), $PHPShopGUI->setSelect('select_action[]', $select_value, 150, true)), array(array(3, 2), array(2, 2)));
-    $Tab2 .= $PHPShopGUI->setField(array('Колонка 11', 'Колонка 12'), array($PHPShopGUI->setSelect('select_action[]', $select_value, 150, true), $PHPShopGUI->setSelect('select_action[]', $select_value, 150, true)), array(array(3, 2), array(2, 2)));
-    $Tab2 .= $PHPShopGUI->setField(array('Колонка 13', 'Колонка 14'), array($PHPShopGUI->setSelect('select_action[]', $select_value, 150, true), $PHPShopGUI->setSelect('select_action[]', $select_value, 150, true)), array(array(3, 2), array(2, 2)));
-    $Tab2 .= $PHPShopGUI->setField(array('Колонка 15', 'Колонка 16'), array($PHPShopGUI->setSelect('select_action[]', $select_value, 150, true), $PHPShopGUI->setSelect('select_action[]', $select_value, 150, true)), array(array(3, 2), array(2, 2)));
-    $Tab2 .= $PHPShopGUI->setField(array('Колонка 17', 'Колонка 18'), array($PHPShopGUI->setSelect('select_action[]', $select_value, 150, true), $PHPShopGUI->setSelect('select_action[]', $select_value, 150, true)), array(array(3, 2), array(2, 2)));
-    $Tab2 .= $PHPShopGUI->setField(array('Колонка 19', 'Колонка 20'), array($PHPShopGUI->setSelect('select_action[]', $select_value, 150, true), $PHPShopGUI->setSelect('select_action[]', $select_value, 150, true)), array(array(3, 2), array(2, 2)));
+    $Tab2 = $PHPShopGUI->setField(array('Колонка 1', 'Колонка 2'), array($PHPShopGUI->setSelect('select_action[]', $select_value, 150, true,false,true), $PHPShopGUI->setSelect('select_action[]', $select_value, 150, true,false,true)), array(array(3, 2), array(2, 2)));
+    $Tab2 .= $PHPShopGUI->setField(array('Колонка 3', 'Колонка 4'), array($PHPShopGUI->setSelect('select_action[]', $select_value, 150, true,false,true), $PHPShopGUI->setSelect('select_action[]', $select_value, 150, true,false,true)), array(array(3, 2), array(2, 2)));
+    $Tab2 .= $PHPShopGUI->setField(array('Колонка 5', 'Колонка 6'), array($PHPShopGUI->setSelect('select_action[]', $select_value, 150, true,false,true), $PHPShopGUI->setSelect('select_action[]', $select_value, 150, true,false,true)), array(array(3, 2), array(2, 2)));
+    $Tab2 .= $PHPShopGUI->setField(array('Колонка 7', 'Колонка 8'), array($PHPShopGUI->setSelect('select_action[]', $select_value, 150, true,false,true), $PHPShopGUI->setSelect('select_action[]', $select_value, 150, true,false,true)), array(array(3, 2), array(2, 2)));
+    $Tab2 .= $PHPShopGUI->setField(array('Колонка 9', 'Колонка 10'), array($PHPShopGUI->setSelect('select_action[]', $select_value, 150, true,false,true), $PHPShopGUI->setSelect('select_action[]', $select_value, 150, true,false,true)), array(array(3, 2), array(2, 2)));
+    $Tab2 .= $PHPShopGUI->setField(array('Колонка 11', 'Колонка 12'), array($PHPShopGUI->setSelect('select_action[]', $select_value, 150, true,false,true), $PHPShopGUI->setSelect('select_action[]', $select_value, 150, true,false,true)), array(array(3, 2), array(2, 2)));
+    $Tab2 .= $PHPShopGUI->setField(array('Колонка 13', 'Колонка 14'), array($PHPShopGUI->setSelect('select_action[]', $select_value, 150, true,false,true), $PHPShopGUI->setSelect('select_action[]', $select_value, 150, true,false,true)), array(array(3, 2), array(2, 2)));
+    $Tab2 .= $PHPShopGUI->setField(array('Колонка 15', 'Колонка 16'), array($PHPShopGUI->setSelect('select_action[]', $select_value, 150, true,false,true), $PHPShopGUI->setSelect('select_action[]', $select_value, 150, true,false,true)), array(array(3, 2), array(2, 2)));
+    $Tab2 .= $PHPShopGUI->setField(array('Колонка 17', 'Колонка 18'), array($PHPShopGUI->setSelect('select_action[]', $select_value, 150, true,false,true), $PHPShopGUI->setSelect('select_action[]', $select_value, 150, true,false,true)), array(array(3, 2), array(2, 2)));
+    $Tab2 .= $PHPShopGUI->setField(array('Колонка 19', 'Колонка 20'), array($PHPShopGUI->setSelect('select_action[]', $select_value, 150, true,false,true), $PHPShopGUI->setSelect('select_action[]', $select_value, 150, true,false,true)), array(array(3, 2), array(2, 2)));
 
     // Закладка 3
     $PHPShopOrm = new PHPShopOrm($GLOBALS['SysValue']['base']['exchanges']);
     $data = $PHPShopOrm->select(array('*'), array('type' => '="import"'), array('order' => 'id DESC'), array("limit" => "1000"));
     $exchanges_value[] = array(__('Создать новую настройку'), 'new');
-    if (is_array($data))
+    if (is_array($data)){
         foreach ($data as $row) {
             $exchanges_value[] = array($row['name'], $row['id']);
             $exchanges_remove_value[] = array($row['name'], $row['id']);
         }
+    }else $exchanges_remove_value=null;
 
     $Tab3 = $PHPShopGUI->setField('Выбрать настройку', $PHPShopGUI->setSelect('exchanges', $exchanges_value, 300, false));
     $Tab3 .= $PHPShopGUI->setField('Сохранить настройку', $PHPShopGUI->setInputArg(array('type' => 'text', 'placeholder' => 'Имя настройки', 'size' => '300', 'name' => 'exchanges_new', 'class' => 'vendor_add')));
+    
+    if(is_array($exchanges_remove_value))
     $Tab3 .= $PHPShopGUI->setField('Удалить настройки', $PHPShopGUI->setSelect('exchanges_remove[]', $exchanges_remove_value, 300, false, false, false, false, 1, true));
 
     // Закладка 4
@@ -1017,11 +1026,21 @@ function actionStart() {
 
     if (empty($_POST['line_limit']))
         $_POST['line_limit'] = 500;
+    
+    if(empty($_POST['bot']))
+        $_POST['bot']=null;
 
     $Tab4 = $PHPShopGUI->setField('Лимит строк', $PHPShopGUI->setInputText(null, 'line_limit', $_POST['line_limit'], 150), 1, 'Задается хостингом');
     $Tab4 .= $PHPShopGUI->setField('Временной интервал', $PHPShopGUI->setInputText(null, 'time_limit', $_POST['time_limit'], 150, __('минут')), 1, 'Задается хостингом');
     $Tab4 .= $PHPShopGUI->setField("Помощник", $PHPShopGUI->setCheckbox('bot', 1, __('Умная загрузка для соблюдения правила ограничений на хостинге'), $_POST['bot'], false, false));
 
+    $Tab1 = $PHPShopGUI->setCollapse('Настройки',$Tab1);
+    $Tab2 = $PHPShopGUI->setCollapse('Подсказка',$PHPShopGUI->setHelp('Если вы загружаете файл, который скачали в меню "База" &rarr; "Экспорт базы", и он содержит штатные заголовки столбцов – сопоставление полей делать <b>не нужно</b>. Если это сторонний прайс со своими названиями колонок, сделайте <b>Cопоставление полей</b>.')).
+            $PHPShopGUI->setCollapse('Сопоставление полей',$Tab2);
+    
+    $Tab3 = $PHPShopGUI->setCollapse('Сохраненные настройки',$Tab3);
+    $Tab4 = $PHPShopGUI->setCollapse('Автоматизация',$Tab4);
+    
     $PHPShopGUI->tab_return = true;
     $PHPShopGUI->setTab(array('Настройки', $Tab1, true), array('Сопоставление полей', $Tab2, true), array('Сохраненные настройки', $Tab3, true), array('Автоматизация', $Tab4, true));
 
@@ -1029,7 +1048,7 @@ function actionStart() {
     $PHPShopModules->setAdmHandler(__FILE__, __FUNCTION__, $data);
 
     // Вывод кнопок сохранить и выход в футер
-    $ContentFooter = $PHPShopGUI->setInput("hidden", "rowID", $data['id'], "right", 70, "", "but") .
+    $ContentFooter = $PHPShopGUI->setInput("hidden", "rowID", true, "right", 70, "", "but") .
             $PHPShopGUI->setInput("submit", "editID", "Сохранить", "right", 70, "", "but", "actionUpdate.exchange.edit") .
             $PHPShopGUI->setInput("submit", "saveID", "Применить", "right", 80, "", "but", "actionSave.exchange.edit");
